Move set up source live os image task to sources
Live OS source will use this task and Live OS handler will use the
source.
